### 2023-08-02 #5729 @editorialbot
[REVIEW]: Paraqus: Exporting Finite Element Simulation Results from Abaqus to VTK
**Submitting author:** @tmfrlm (<a href="http://orcid.org/0000-0002-3337-5891">Tim Furlan</a>)
**Repository:** https://github.com/instituteofmechanics/paraqus
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v1.0.5
