### 2023-09-05 #5729 @prashjha
@KParas, good point. You can ask the author, @tmfrln, for clarification. While the "Contributions" section is not mandatory, as a reviewer, you can request authors to have one in the draft.
