### 2023-09-06 #5729 @alizma
Hi @tmfrln, could you add a CONTRIBUTING.md file that discusses "guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"?

Also I know the dependencies are minimal, but nevertheless, these should be listed somewhere in the Installation section of your documentation, and ideally also provided as a ``requirements.txt`` or equivalent for completeness sake.

Also @prashjha, could you suggest how I should confirm end-to-end functionality of the provided examples if I do not have access to Abaqus?
