### 2023-09-07 #5729 @tmfrln
Hi @alizma,

- I added the file that you requested to the repository.

- The dependencies are indeed minimal, only numpy is required to run Paraqus. This dependency is specifies in the file pyproject.toml (without a version number, since we are not aware of a specific requirement). The python interpreter shipped with Abaqus has numpy preinstalled, and if Paraqus is installed standalone via pip, numpy should be automatically installed also. I updated the documentation with this information as well, but think that a requirements.txt would conflict with the pyproject.toml.
