### 2023-09-12 #5729 @prashjha
> Hi @tmfrln, could you add a CONTRIBUTING.md file that discusses "guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"?
>
> Also I know the dependencies are minimal, but nevertheless, these should be listed somewhere in the Installation section of your documentation, and ideally also provided as a `requirements.txt` or equivalent for completeness sake.
>
> Also @prashjha, could you suggest how I should confirm end-to-end functionality of the provided examples if I do not have access to Abaqus?

@alizma, it looks like the library comes with the Abaqus-produced files to test the scripts. You do not need to run Abaqus to test this library. The authors do provide instructions to reproduce the Abaqus output file in case anyone is interested. But for the review, you can assume that the Abaqus output file is given and the scripts work as intended. @tmfrln, does it sound right?
