### 2023-09-17 #5729 @tmfrln
@prashjha The library is split up in two parts: The vtk writers and a general data model are independent of Abaqus and can be used/tested in any python environment. The second part is supposed to be run in Abaqus and implements the interface of the Abaqus database to the data model. So for some of the examples, Abaqus in indeed required to test them.
One of the main ideas of this package is to export results from Abaqus once, and have them in a format that can be shared with people who do not use Abaqus afterwards.
If this would be helpful, i could offer to showcase the Abaqus-related examples on a zoom call or similar. I am however not sure if that is ok with JOSS's requirements for peer review. If the group of @KParas has access to Abaqus, maybe it would be an option for him to share if the examples worked for him?
