### 2024-09-18 #5729 @prashjha
Howdy, @Huzaifg, @lelaus, @Brookluo, @jannisteunissen, would you be interested in reviewing the JOSS submission above ([pdf link](https://github.com/openjournals/joss-papers/blob/joss.05729/joss.05729/10.21105.joss.05729.pdf))? The corresponding software is at this [link](https://github.com/tmfrln/paraqus).

This [link](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) covers almost everything about the review process; I will be happy to help with more information.

Please let me know if you are interested. Thank you!
