### 2024-09-19 #5729 @jannisteunissen
@prashjha I might not be the best reviewer since I have never used abaqus and don't have access to it
