### 2024-09-19 #5729 @Huzaifg
Hello Prashant,

I don't currently have access to Abaqus. Can I do the review without it? If
yes, I would be happy to review

Best
Huzaifa

On Wed, Sep 18, 2024, 3:43 PM Prashant K. Jha ***@***.***>
wrote:

> Hi @jstollberg <https://github.com/jstollberg>, the delay is due to two
> reviewers ghosting us completely. I will try to find one more reviewer, and
> based on their assessment and @alizma <https://github.com/alizma>'s
> partial assessment, it should be sufficient to make a decision.
>
