### 2024-09-19 #5729 @prashjha
Thank you for your positive response, @Huzaifg. Let me confirm with the authors.

Hi @tmfrln and @jstollberg, please let us know if the review of your repository is possible without access to Abaqus.
