### 2024-09-19 #5729 @lelaus
Hello @prashjha ,

Thank you for contacting me.
I would be very happy to participate in this review.

However, I have never used Abaqus and have no licence to use it at the moment.

In your opinion, is it possible to do a full review despite these limitations?
In case of a positive answer, I would be available.

I look forward to your reply.
