### 2024-09-22 #5729 @tmfrln
Hi everyone,

first of all, thanks @lelaus and @Huzaifg for your interest in taking part in the review.

I think a complete review is not possible without access to Abaqus. However, there is a learning edition that can be downloaded for free [here](https://www.3ds.com/de/edu/education/students/solutions/abaqus-le), but is only available on Windows. A registration on the Simulia site may also be required.

With this edition, reviewers can run the complete test suite for paraqus, but only some of the examples:
example_abaqus_aluminum_bending.py, example_abaqus_rivet_forming.py. and example_abaqus_extrusion.py will not work with the learning version.

From my point of view, the remaining tests and examples should be enough to review the package though.

P.S.: Since this has been up for review for quite some time now, i would like to kindly ask reviewers to only participate if they can review the package in the near future - thanks!
