### 2024-10-21 #5729 @lelaus
Hi @prashjha and @tmfrln, thanks for adding me as reviewer. I'll do my best. Before start, I just want to advice you that unfortunately my review will not be based on the usage of the learning version of Abaqus since I have only Linux Debian OS.
