### 2024-12-06 #5729 @tmfrln
Dear @lelaus, we would like to thank you again for the thorough review and the suggestions. We have included most of them in the repo, and where we did not we tried to supply our reasoning in the issue comments. As far as i am aware, only a citation for Paraqus is missing now, which is what we hope to include after a DOI is available from JOSS.

Please be so kind to check whether your comments are addressed in a satisfactory manner. We would also be happy to include an acknowledgement for them - if you would like to be mentioned in such a way, please let me know by what name we shoudl reference you.

@prashjha Since we did only get one real review yet, is it possible to move forward with the submission after @lelaus confirms that we addressed the issues?
