### 2024-12-15 #5729 @lelaus
Hi @prashjha ,

I've finished my part of the review. I made several comments and suggestions about the paper, the software, and its documentation by opening specific issues in the *Paraqus* repo. The authors/developers have dealt with these satisfactorily, either by implementing changes/additions or by providing appropriate justifications.

I think that the *Paraqus* package will be useful for managing the sharing of FEM (and not only) simulation results generated by commercial software tools with a wider scientific community by converting them to the open-source VTK format. In my opinion, the package is well designed, implemented, and documented, and the paper is concise and well written. Unfortunately, I don't have access to the ABAQUS software, so I couldn't verify that the tests and examples focusing on ABAQUS functionalities work properly. This is a point that may need further verification.

From my side, I support the publication of *Paraqus* in JOSS.

I would also like to thank you, @tmfrln , for asking me to be mentioned in the *Acknowledgments* section. I would be very honored by such a mention, and there would be no problem in sharing my name with you. But on this point I will leave it to @prashjha to say whether this mention is possible according to the review guidelines of JOSS.
