### 2024-12-19 #5729 @tmfrln
Hi @lelaus,

the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#review-process) explicitly encourage acknowledgement of reviewer contributions:

> Reviewer and editor contributions, like any other contributions, should be acknowledged in the repository.

We can of course still wait for @prashjha to confirm this.
