### 2025-01-06 #5729 @tmfrln
Dear @prashjha,
since @lelaus finished their review 3 weeks ago, could you share an estimate when a decision regarding the publication can be expected? This would be helpful for us to plan if/how we can cite paraqus in an upcoming publication.
