### 2025-01-17 #5729 @prashjha
Hi @tmfrln, based on @lelaus reviews and old reviews by @alizma (no longer responsive to our pings), I am happy to recommend acceptance of your submission.

We have a few things to take care of:
1. Let me know if you have updated version of the software.
2. Please create a zenodo archive of your software with the version you want to associated to your JOSS article submission. Make sure that the title of zenodo archive matches the title of your JOSS paper. Once done, share the archive DOI.
3. I am reading your draft one final time and let you know if I have any suggestions.
