### 2025-01-17 #5729 @prashjha
@tmfrln, my only suggestion for the JOSS paper is that it will be nice to discuss how your library is used together with Abaques. Maybe provide a syntax and example in the paper.
