### 2025-01-18 #5729 @tmfrln
@prashjha Thank you for your input, we will include a small syntax example in the paper and let you know when the repository is ready.

@lelaus Please let me know the name under which you would like to be acknowledged. If you would prefer not to post it here, you can also email me at tim.furlan@tu-dortmund.de
