### 2025-01-21 #5729 @tmfrln
Hi @prashjha,

we updated the paper according to your recommendation, including a syntax example. We also added the acknowledgement of @lelaus. During the review, we made some smaller changes in the code. The latest version is v1.0.5.

Furthermore, we moved the repository to the github account of my organisation at https://github.com/instituteofmechanics/paraqus (but for now it is still available at the old repo also).

The zenodo archive is located at https://zenodo.org/records/14710993 with DOI 10.5281/zenodo.14710993

Please let me know if I have missed any other information that is required.
