### 2025-01-22 #5729 @tmfrln
> @tmfrln Do you want me to associate this repo [https://github.com/instituteofmechanics/paraqus](https://github.com/instituteofmechanics/paraqus) to your JOSS paper?

That would be perfect, thank you!
