+++ paper.md
@@ -39,7 +39,7 @@ Abaqus is an example of commercial FE software that is widely used in the academ
 - the data can be shared with other researchers and users without access to expensive software, and 
 - specialized open source software like ParaView can be used to create visualizations that would be tedious or impossible to generate in the post-processing module of Abaqus.
 
-Some software packages similar to Paraqus have been released over the last years. However, these tools suffer from great limitations and are, therefore, not as versatile and straightforward as Paraqus. A script to export Abaqus results to VTK format was published in @odb2vtk under the name ``odb2vtk``. This script offers limited options to customize the export data, and supports only a subset of the finite elements available in Abaqus. *AbaPy* is a Python package that is not aimed solely at post-processing, but also at the automatic creation of Abaqus models and run of Abaqus simluations [@abapy]. Although *AbaPy* offers the option to export field outputs to VTK format, the wider range of applications comes with increased complexity in terms of both implementation and usage. Paraqus falls between these existing options: it is organized as a package with a small and intuitive API, but allows a wide range of customization of exports, and writes the more efficient binary version of the VTK format. It also provides options to group exports from multiple simulations, time steps, or bodies, using ParaView Data (PVD) files. Furthermore, Paraqus performs exports noticeably faster than *AbaPy*, especially in case of large models.
+Some software packages similar to Paraqus have been released over the last years. However, these tools suffer from great limitations and are, therefore, not as versatile and straightforward as Paraqus. A script to export Abaqus results to VTK format was published in @odb2vtk under the name ``odb2vtk``. This script offers limited options to customize the export data, and supports only a subset of the finite elements available in Abaqus. *AbaPy* is a Python package that is not aimed solely at post-processing, but also at the automatic creation of Abaqus models and run of Abaqus simluations [@abapy]. Although *AbaPy* offers the option to export field outputs to VTK format, the wider range of applications comes with increased complexity in terms of both implementation and usage. Paraqus falls between these existing options: it is organized as a package with a small and intuitive API, but allows a wide range of customization of exports, and writes the more efficient binary version of the VTK format. It also provides options to group exports from multiple simulations, time steps, or bodies, using ParaView data (PVD) files. Furthermore, Paraqus performs exports noticeably faster than *AbaPy*, especially in case of large models.
 
 # Syntax example
 
