### 2025-01-28 #5729 @tmfrln
> [@tmfrln](https://github.com/tmfrln), we are almost there. Please update the title of your zenodo https://zenodo.org/records/14710993 archive to match the title of JOSS paper.

Thank you - i just updated the title there.
