### 2025-02-04 #5729 @kyleniemeyer
Hi @tmfrln, it looks like a list in the first paragraph of the statement of need section is not formatted properly in the paper. Can you please fix this?

Also, the paper mentions Abaqus, but I don't see any citations/references/links for it. Are there any appropriate that could be added?
