### 2025-02-04 6f6251f
adjust list formatting for joss engine
