+++ paper.bib
@@ -1,4 +1,3 @@
-
 @article{odb2vtk,
 title = {ParaView visualization of {A}baqus output on the mechanical deformation of complex microstructures},
 journal = {Computers & Geosciences},
@@ -15,4 +14,17 @@ doi = {10.5281/zenodo.50550},
 publisher = {Zenodo}, 
 author = {Ludovic Charleux and Laurent Bizet and Vincent Keryvin and mousta}, 
 year = {2016}
+}
+
+@incollection{abaqus,
+title = {ABAQUS},
+editor = {Ove Stephansson and Lanru Jing and Chin-Fu Tsang},
+series = {Developments in {G}eotechnical {E}ngineering},
+publisher = {Elsevier},
+volume = {79},
+pages = {565--570},
+year = {1996},
+booktitle = {Coupled {T}hermo-{H}ydro-{M}echanical {P}rocesses of {F}ractured {M}edia},
+doi = {10.1016/S0165-1250(96)80047-2},
+author = {L. B{\"o}rgesson}
 }
