+++ README.md
@@ -2,7 +2,7 @@
 
 ## Introduction
 
-Paraqus is a Python package to convert simulation results, for example from finite element or CFD simulations, to the widely used and open *Visualization Toolkit* (VTK) file format. It was initially designed for the export of FE results from the commercial FE software [Abaqus](https://www.3ds.com/products/simulia/abaqus), but its modular structure allows it to be extended to other software. VTK files can be visualized for example by the free and open source software [ParaView](https://www.paraview.org/>). Paraqus is intended for anyone who wants to convert simulation results, especially from Abaqus, to a non-proprietary format, either to use tools such as ParaView for their own visualizations, or to exchange data in an open format.
+Paraqus is a Python package to convert simulation results, for example from finite element or CFD simulations, to the widely used and open *Visualization Toolkit* (VTK) file format. It was initially designed for the export of FE results from the commercial FE software [Abaqus](https://www.3ds.com/products/simulia/abaqus), but its modular structure allows it to be extended to other software. VTK files can be visualized for example by the free and open source software [ParaView](https://www.paraview.org). Paraqus is intended for anyone who wants to convert simulation results, especially from Abaqus, to a non-proprietary format, either to use tools such as ParaView for their own visualizations, or to exchange data in an open format.
 
 
 ## Documentation
