### 2025-02-07 10ef5e6
removed very old abaqus paper
