### 2025-02-07 #5729 @tmfrln
Hi @kyleniemeyer,

we fixed the table formatting, thanks for the heads up.

As for Abaqus, there is no scientific citation for the software (i.e. no DOI). A link to the website (https://www.3ds.com/products/simulia/abaqus) could be added if you think this would add value, but we originally omitted this to avoid the link going dead if the product website changes in the future. If still prefer to have a link i can of course add one.

Is a new release in the repo required after these changes to the paper? (Source code was not affected)
